use py rocal tensor class
